requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
sqlite3
joblib==1.3.1
APScheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.0.0
Werkzeug==2.3.7